SimpleSimulationUpdate
I'm excited to announce that I've successfully simulated a normal environment for a simple computer shop! The program includes several main features, such as the ability to show all computers, view whether they are online or offline, and see relevant information about each computer.

One of the most exciting things about this update is that users no longer need to log in to an account to use the simulation program. This means that everyone can take advantage of the program's features and get a feel for how a computer shop operates.